+++ Procfile
@@ -1,0 +1,2 @@
+web: NODE_ENV=production PORT=3000 STATIC_DIR=client DATA_DIR=data node index.js
+
