+++ client/index.html
@@ -12,6 +12,7 @@
   </head>
 
   <body>
+    <p>hello world</p>
     <script>
       console.log('hello');
     </script>
