+++ package.json
@@ -41,7 +41,7 @@
     "html-validate": "^5.2.0",
     "jest": "^25.1.0",
     "jsdoc-to-markdown": "^7.0.1",
-    "nodemon": "^2.0.2",
+    "nodemon": "^2.0.13",
     "prettier": "^2.3.2",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
